libcontainer: init: only pass stateDirFd when creating a container
If we pass a file descriptor to the host filesystem while joining a
container, there is a race condition where a process inside the
container can ptrace(2) the joining process and stop it from closing its
file descriptor to the stateDirFd. Then the process can access the
*host* filesystem from that file descriptor. This was fixed in part by
5d93fed3d27f ("Set init processes as non-dumpable"), but that fix is
more of a hail-mary than an actual fix for the underlying issue.

To fix this, don't open or pass the stateDirFd to the init process
unless we're creating a new container. A proper fix for this would be to
remove the need for even passing around directory file descriptors
(which are quite dangerous in the context of mount namespaces).

There is still an issue with containers that have CAP_SYS_PTRACE and are
using the setns(2)-style of joining a container namespace. Currently I'm
not really sure how to fix it without rampant layer violation.

[@kolyshkin: backport to docker-1.13.1-rhel branch]

Fixes: CVE-2016-9962
Fixes: 5d93fed3d27f ("Set init processes as non-dumpable")
Signed-off-by: Aleksa Sarai <asarai@suse.de>
(cherry picked from commit e034cedce7df632eaa52636d34d1d8364733b402)
Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>