init: switch away from stateDirFd entirely
While we have significant protections in place against CVE-2016-9962, we
still were holding onto a file descriptor that referenced the host
filesystem. This meant that in certain scenarios it was still possible
for a semi-privileged container to gain access to the host filesystem
(if they had CAP_SYS_PTRACE).

Instead, open the FIFO itself using a O_PATH. This allows us to
reference the FIFO directly without providing the ability for
directory-level access. When opening the FIFO inside the init process,
open it through procfs to re-open the actual FIFO (this is currently the
only supported way to open such a file descriptor).

[@kolyshkin: backport to docker-1.13.1-rhel branch]

Signed-off-by: Aleksa Sarai <asarai@suse.de>
(cherry picked from commit 7d66aab77abfdd8e2893eb70647fd941f26091b1)
Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>