libctr/init_linux: reorder chdir
commit 5e0e67d76cc99d76c8228d48f38f37034503f315 moved the chdir to be one of the
first steps of finalizing the namespace of the container.

However, this causes issues when the cwd is not accessible by the user running runc, but rather
as the container user.

Thus, setupUser has to happen before we call chdir. setupUser still happens before setting the caps,
so the user should be privileged enough to mitigate the issues fixed in 5e0e67d76cc99d76c8228d48f38f37034503f315

Signed-off-by: Peter Hunt <pehunt@redhat.com>