libct/cg/fs/freezer: make sure to thaw on failure
Function (*FreezerGroup).Set has a few paths where in can return an
error. In any case, if an error is returned, we failed to freeze,
and we need to thaw to avoid leaving the cgroup in a stuck state.

Cherry-picked: fcd7fe85e1ea706fbdfb383824e9f390a53a34e9

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>