libct/cg/fscommon.OpenFile: reverse checks order
In case openat2() is not available, it does not make sense to calculate
relpath (and check if path has /sys/fs/cgroup prefix).

Reverse the order of checks to not do that in case openat2 is not
available.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>
(cherry picked from commit 272d34f5e82b7a0dd354fa828e65e218241080ad)
Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>