init: don't special-case logrus fds
We close the logfd before execve so there's no need to special case it.
In addition, it turns out that (*os.File).Fd() doesn't handle the case
where the file was closed and so it seems suspect to use that kind of
check.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>
(cherry picked from commit 25862d58dabbad8f928d51377c3d0041cfc9f550)
Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>